feat: parse/replace images in post and detect thumbnail
Images are assumed to be Obsidian format embeds - ![[<image.png>]] - these are replaced with <img> tags pointing to the DOMAIN env var. On this domain, the express server statically serves the images contained in all the posts from the folder IMAGES_PATH (this should be set to something like the attachments folder for Obsidian, or some other folder containing all images for all posts).